fs/stream: limit unix read/write to 2GB at a time
On some platforms, notably OSX, read/readv/write/writev all return
an EINVAL if the requested number of bytes is greater than INT32_MAX.

We therefore limit each such call to 2GB on all platforms. On
platforms where this is not an issue, the overhead of an extra syscall
for every 2GB of read/write is negligible.

+++ src/uv-common.h
@@ -188,7 +188,7 @@ void uv__work_submit(uv_loop_t* loop,
 
 void uv__work_done(uv_async_t* handle);
 
-size_t uv__count_bufs(const uv_buf_t bufs[], unsigned int nbufs);
+uint64_t uv__count_bufs(const uv_buf_t bufs[], unsigned int nbufs);
 
 int uv__socket_sockopt(uv_handle_t* handle, int optname, int* value);
